pyproject.toml
[tool.poetry]
name = "fractions"
version = "0.1.0"
description = ""
authors = ["Rhett D. Hudson, Ph.D. <rhett.hudson@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
solidpython = "^1.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
